Functions.java: substract()-method
Added code to substract()-method. the method is given two parameters:  originalNumber and numbers. originalNumber is a double and numbers is an array of doubles. The length of the array can vary. The method uses for-loop to substract all the numbers from originalNumber. Added a couple of tests for substract()-method in the main()-method.